docker_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
docker==7.0.0
python-dotenv==1.0.1
loguru==0.7.2
psutil==5.9.8
websockets==12.0
pydantic==2.7.4
